+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.4
+* Android Send Cancel Event [235](https://github.com/eopeter/flutter_mapbox_navigation/pull/235)
+* iOS Receive Feedback Sent to Mapbox on Dart Side; Ability to Turn On/Off Show Feedback [235](https://github.com/eopeter/flutter_mapbox_navigation/pull/235)
+* Add Free Drive Mode [240](https://github.com/eopeter/flutter_mapbox_navigation/pull/240)
+
 ## 0.1.3
 * Android Send Cancel Event [236](https://github.com/eopeter/flutter_mapbox_navigation/pull/236)